Pull getBaseImageName into a util function
